remove duplicate restricted middleware
